requirements.txt
# Core Framework
fastapi==0.116.1
uvicorn==0.41.0
starlette==0.47.3
httpx==0.28.1
requests==2.32.5
aiohttp==3.13.3

# Database
sqlalchemy==2.0.43
alembic==1.16.5
aiosqlite==0.21.0
asyncpg==0.31.0
pgvector==0.4.2

# AI/LLM Providers
openai==2.21.0
anthropic==0.83.0
google-genai==1.64.0
groq==1.0.0
cohere==5.20.6
mistralai==1.12.4
xai-sdk==1.7.0
ollama==0.6.1

# Pydantic AI
pydantic==2.12.5
pydantic-ai==1.62.0
pydantic-ai-slim==1.62.0
pydantic-core==2.41.5
pydantic-settings==2.13.1
pydantic-graph==1.62.0
pydantic-evals==1.62.0
pydantic-extra-types==2.11.0
annotated-types==0.7.0
typing-extensions==4.15.0
typing-inspection==0.4.2

# Vector Stores & Embeddings
chromadb==0.6.3
chroma-hnswlib==0.7.6
faiss-cpu==1.13.2
qdrant-client==1.17.0
sentence-transformers==5.2.3
transformers==5.2.0
tokenizers==0.22.2
safetensors==0.7.0

# Document Processing
python-docx==1.2.0
pypdf2==3.0.1
pypdf==4.3.1
markdown==3.10.2
beautifulsoup4==4.14.3
soupsieve==2.8.3
python-multipart==0.0.22
pillow==12.1.1
pytesseract==0.3.13
tesseract==0.1.3
lxml==6.0.2

# Data Science & ML
numpy==2.2.6
pandas==2.3.3
scikit-learn==1.8.0
scipy==1.17.0
torch==2.8.0
torchvision==0.23.0
torchaudio==2.8.0
accelerate==0.34.2
sentencepiece==0.2.1
joblib==1.5.3
threadpoolctl==3.6.0
tqdm==4.67.3
networkx==3.6.1
sympy==1.14.0
mpmath==1.3.0
dill==0.4.1
regex==2026.1.15
filelock==3.24.2
fsspec==2026.2.0
packaging==25.0
pyyaml==6.0.3

# LLM Serving & Optimization
vllm==0.11.0
ray==2.54.0
outlines-core==0.2.11
lm-format-enforcer==0.11.3
compressed-tensors==0.11.0
gguf==0.17.1
einops==0.8.2

# Streamlit & UI
streamlit==1.54.0
altair==6.0.0
pydeck==0.9.1
jinja2==3.1.6
markupsafe==3.0.2
blinker==1.9.0
watchdog==6.0.0
pyarrow==23.0.1
tornado==6.5.4

# AWS & Cloud
boto3==1.42.54
botocore==1.42.54
s3transfer==0.16.0
jmespath==1.1.0

# Authentication & Security
cryptography==46.0.5
pyjwt==2.11.0
oauthlib==3.3.1
requests-oauthlib==2.0.0
authlib==1.6.8
bcrypt==5.0.0
rsa==4.9.1
pyasn1==0.6.2
pyasn1-modules==0.4.2
google-auth==2.48.0
cachetools==6.2.6

# Monitoring & Observability
loguru==0.7.3
opentelemetry-api==1.39.1
opentelemetry-sdk==1.39.1
opentelemetry-proto==1.39.1
opentelemetry-exporter-otlp-proto-common==1.39.1
opentelemetry-exporter-otlp-proto-grpc==1.39.1
opentelemetry-exporter-otlp-proto-http==1.39.1
opentelemetry-instrumentation==0.60b1
opentelemetry-instrumentation-asgi==0.60b1
opentelemetry-instrumentation-fastapi==0.60b1
opentelemetry-instrumentation-httpx==0.60b1
opentelemetry-semantic-conventions==0.60b1
opentelemetry-util-http==0.60b1
prometheus-fastapi-instrumentator==7.1.0
prometheus-client==0.24.1
sentry-sdk==2.53.0
logfire==4.25.0
logfire-api==4.25.0
posthog==7.9.3
python-json-logger==4.0.0

# Utilities
python-dotenv==1.1.1
tenacity==8.5.0
pydantic-settings==2.13.1
python-dateutil==2.9.0.post0
pytz==2025.2
tzdata==2025.3
distro==1.9.0
certifi==2026.1.4
charset-normalizer==3.4.4
idna==3.10
urllib3==2.6.3
toml==0.10.2
jsonschema==4.26.0
jsonschema-specifications==2025.9.1
referencing==0.37.0
rpds-py==0.30.0
wrapt==1.17.3
greenlet==3.2.4
click==8.1.8
typer==0.23.1
typer-slim==0.23.1
shellingham==1.5.4
rich==14.3.2
rich-rst==1.3.2
rich-toolkit==0.19.4
markdown-it-py==4.0.0
mdurl==0.1.2
pygments==2.19.2
platformdirs==4.9.2
pathspec==1.0.4
argcomplete==3.6.3
importlib-metadata==8.7.1
importlib-resources==6.5.2
zipp==3.23.0
more-itertools==10.8.0
jaraco-classes==3.4.0
jaraco-context==6.1.0
jaraco-functools==4.4.0
keyring==25.7.0
docutils==0.22.4
invoke==2.2.1
cyclopts==4.5.4
overrides==7.7.0

# Async & Performance
asgiref==3.11.1
anyio==4.10.0
sniffio==1.3.1
watchfiles==1.1.1
httptools==0.7.1
websockets==15.0.1
aiofiles==24.1.0
aiofile==3.9.0
caio==0.9.25
multidict==6.7.1
yarl==1.22.0
frozenlist==1.8.0
aiosignal==1.4.0
propcache==0.4.1
h11==0.16.0
httpcore==1.0.9
h2==4.3.0
hpack==4.1.0
hyperframe==6.1.0
aiohappyeyeballs==2.6.1

# MCP & AI Tools
mcp==1.26.0
fastmcp==3.0.1
genai-prices==0.0.54
docstring-parser==0.17.0
eval-type-backport==0.3.1
annotated-doc==0.0.4
beartype==0.22.9
griffelib==2.0.0
interegular==0.3.3
lark==1.2.2
partial-json-parser==0.2.1.1.post7
openai-harmony==0.0.8

# Serialization
orjson==3.11.7
msgspec==0.20.0
msgpack==1.1.2
protobuf==6.33.5
flatbuffers==25.12.19
fastavro==1.12.1
cbor2==5.8.0
pybase64==1.4.3
jiter==0.13.0
jsonref==1.1.0
openapi-pydantic==0.5.1
pathable==0.5.0
jsonschema-path==0.4.1

# Git & Version Control
gitpython==3.1.46
gitdb==4.0.12
smmap==5.0.2

# System
psutil==7.2.2
setproctitle==1.3.7
py-cpuinfo==9.0.0
portalocker==3.2.0
diskcache==5.6.3
frozendict==2.4.7
mmh3==5.2.0
pyzmq==27.1.0

# Windows specific
pywin32==311 ; sys_platform == 'win32'
pywin32-ctypes==0.2.3 ; sys_platform == 'win32'
win32-setctime==1.2.0 ; sys_platform == 'win32'
colorama==0.4.6 ; sys_platform == 'win32'

# Networking
dnspython==2.8.0
email-validator==2.3.0
httpcore==1.0.9
httpx-sse==0.4.3
sse-starlette==3.0.3
websocket-client==1.9.0
kubernetes==35.0.0
py-key-value-aio==0.4.4
temporalio==1.20.0
nexus-rpc==1.2.0
grpcio==1.78.1

# Development & Testing (optional - can be moved to requirements-dev.txt)
pytest==8.4.2
pytest-asyncio==0.26.0
pytest-cov==4.1.0
black==23.12.1
mypy==1.19.1
mypy-extensions==1.1.0
isort==5.13.2
ruff==0.3.7
coverage==7.13.4
build==1.4.0
pyproject-hooks==1.2.0
iniconfig==2.3.0
pluggy==1.6.0
exceptiongroup==1.3.0
executing==2.2.1
prompt-toolkit==3.0.52
wcwidth==0.6.0
pyperclip==1.11.0

# Additional ML/DL
numba==0.61.2
llvmlite==0.44.0
onnxruntime==1.24.2
opencv-python-headless==4.11.0.86
soundfile==0.13.1
soxr==1.0.0
cupy-cuda12x==13.6.0
fastrlock==0.8.3
cloudpickle==3.1.2
depyf==0.19.0
librt==0.8.1
ninja==1.13.0
mistral-common==1.9.1

# HuggingFace
huggingface-hub==1.4.1
hf-xet==1.2.0
rignore==0.7.6
sentence-transformers==5.2.3
transformers==5.2.0

# Utilities (additional)
ag-ui-protocol==0.1.13
astor==0.8.1
backoff==2.2.1
blake3==1.0.8
cffi==2.0.0
durationpy==0.10
fastapi-cli==0.0.23
fastapi-cloud-cli==0.13.0
fastar==0.8.0
googleapis-common-protos==1.72.0
pycountry==26.2.16
pypika==0.51.1
types-protobuf==6.32.1.20260221
types-requests==2.32.4.20260107
setuptools==79.0.1
six==1.17.0
pycparser==3.0